Add a hack to the playwindow log bits saving
Add a hack to the playwindow log bits saving. The user property value
turns into a qlonglong at some point (possibly when reading it back)
and the playwindow writes a quint. So this means that the property
values no longer actually match when the property is synced over IPC
back to the editor host from the game. This is a hack fix converting
from qint to qlong (qint64) trying to make the types match.

+++ editor/gui/playwindow.cpp
@@ -1012,7 +1012,7 @@ void PlayWindow::SaveState(const QString& key_prefix)
     mWorkspace.SetUserProperty(key_prefix + "_show_eventlog", mUI.dockWidget->isVisible());
     mWorkspace.SetUserProperty(key_prefix + "_geometry", saveGeometry());
     mWorkspace.SetUserProperty(key_prefix + "_toolbar_and_dock_state", saveState());
-    mWorkspace.SetUserProperty(key_prefix + "_log_bits", mAppEventLog.GetShowBits());
+    mWorkspace.SetUserProperty(key_prefix + "_log_bits", (qint64)mAppEventLog.GetShowBits());
     mWorkspace.SetUserProperty(key_prefix + "_debug_draw", (bool)GetValue(mUI.actionToggleDebugDraw));
     mWorkspace.SetUserProperty(key_prefix + "_debug_log", (bool)GetValue(mUI.actionToggleDebugLog));
     mWorkspace.SetUserProperty(key_prefix + "_debug_msg", (bool)GetValue(mUI.actionToggleDebugMsg));
